refactor: 会员模块和营销模块重命名

+++ mall-product/product-boot/src/main/java/com/youlai/mall/product/model/entity/Spec.java
@@ -30,9 +30,9 @@ public class Spec implements Serializable {
     private String name;
 
     /**
-     * 输入方式（1：手动输入，2：列表选择）
+     * 输入方式：1->手动输入，2->列表选择
      */
-    private Byte inputType;
+    private Integer inputType;
 
     /**
      * 可选值列表（以逗号分隔，仅当输入方式为2时使用）
@@ -62,5 +62,5 @@ public class Spec implements Serializable {
     /**
      * 逻辑删除标识（0：未删除，1：已删除）
      */
-    private Byte isDeleted;
+    private Integer isDeleted;
 }
